requirements.txt
aerich==0.5.8
aiomysql==0.0.21
aiosqlite==0.17.0
anyio==3.3.2
asgiref==3.4.1
bcrypt==3.2.0
cffi==1.14.6
click==8.0.2
colorama==0.4.4
ddlparse==1.10.0
dictdiffer==0.9.0
dnspython==2.1.0
email-validator==1.1.3
fastapi==0.68.2
fastapi-crudrouter==0.8.4
fastapi-pagination==0.9.0
fastapi-users==8.1.0
fastapi-users-db-tortoise==1.0.0
h11==0.12.0
idna==3.2
iso8601==0.1.16
itsdangerous==2.0.1
makefun==1.11.3
passlib==1.7.4
pycparser==2.20
pydantic==1.8.2
PyJWT==2.1.0
PyMySQL==0.9.3
pyparsing==2.4.7
pypika-tortoise==0.1.1
python-multipart==0.0.5
pytz==2021.3
six==1.16.0
sniffio==1.2.0
starlette==0.14.2
starlette-context==0.3.3
tortoise-orm==0.17.8
typing-extensions==3.10.0.2
uvicorn==0.15.0
